+++ src/test/java/es/in2/issuer/backend/shared/domain/policy/rules/RequireOrganizationRuleTest.java
@@ -13,7 +13,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_succeedsWhenOrganizationMatches() {
-        PolicyContext ctx = new PolicyContext("ORG-123", List.of(), null, null, null, false, false, null, null);
+        PolicyContext ctx = new PolicyContext("ORG-123", List.of(), null, null, null, false, false, null, null, null);
 
         StepVerifier.create(rule.evaluate(ctx, "ORG-123"))
                 .verifyComplete();
@@ -21,7 +21,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_succeedsWhenSysAdmin() {
-        PolicyContext ctx = new PolicyContext("ADMIN_ORG", List.of(), null, null, null, true, false, null, null);
+        PolicyContext ctx = new PolicyContext("ADMIN_ORG", List.of(), null, null, null, true, false, null, null, null);
 
         StepVerifier.create(rule.evaluate(ctx, "COMPLETELY_DIFFERENT_ORG"))
                 .verifyComplete();
@@ -29,7 +29,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_failsWhenOrganizationDoesNotMatchAndNotSysAdmin() {
-        PolicyContext ctx = new PolicyContext("ORG-123", List.of(), null, null, null, false, false, null, null);
+        PolicyContext ctx = new PolicyContext("ORG-123", List.of(), null, null, null, false, false, null, null, null);
 
         StepVerifier.create(rule.evaluate(ctx, "ORG-456"))
                 .expectErrorMatches(e ->
@@ -40,7 +40,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_sysAdminBypassesOrgCheck() {
-        PolicyContext ctx = new PolicyContext("ADMIN_ORG", List.of(), null, null, null, true, false, null, null);
+        PolicyContext ctx = new PolicyContext("ADMIN_ORG", List.of(), null, null, null, true, false, null, null, null);
 
         // Even though orgs don't match, sys-admin bypasses the check
         StepVerifier.create(rule.evaluate(ctx, "ANY_ORG"))
@@ -49,7 +49,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_tenantAdminBypassesOrgCheck() {
-        PolicyContext ctx = new PolicyContext("TENANT_ADMIN_ORG", List.of(), null, null, null, false, true, null, null);
+        PolicyContext ctx = new PolicyContext("TENANT_ADMIN_ORG", List.of(), null, null, null, false, true, null, null, null);
 
         StepVerifier.create(rule.evaluate(ctx, "DIFFERENT_ORG"))
                 .verifyComplete();
@@ -57,7 +57,7 @@ class RequireOrganizationRuleTest {
 
     @Test
     void evaluate_failsWhenNotSysAdminNotTenantAdminAndOrgMismatch() {
-        PolicyContext ctx = new PolicyContext("ORG-A", List.of(), null, null, null, false, false, null, null);
+        PolicyContext ctx = new PolicyContext("ORG-A", List.of(), null, null, null, false, false, null, null, null);
 
         StepVerifier.create(rule.evaluate(ctx, "ORG-B"))
                 .expectErrorMatches(e ->